requirements.txt
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
croniter==1.2.0
cryptography==36.0.1
dask==2022.1.1
distributed==2022.1.1
docker==4.2.2
dropbox==11.27.0
fsspec==2022.1.0
greenlet==1.1.2
HeapDict==1.0.1
idna==3.3
Jinja2==3.0.3
locket==0.2.1
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-oneofschema==3.0.1
msgpack==1.0.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
packaging==21.3
partd==1.2.0
pdfminer.six==20211012
pdfplumber==0.6.0
pendulum==2.1.2
Pillow==9.0.1
ply==3.11
prefect==0.15.13
protobuf==3.19.4
psutil==5.9.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
pypiwin32==223
python-box==5.4.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
pytz==2021.3
pytzdata==2020.1
pywin32==303
PyYAML==6.0
requests==2.27.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
stone==3.3.1
tabulate==0.8.9
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.2
tornado==6.1
urllib3==1.26.8
Wand==0.6.7
websocket-client==1.2.3
zict==2.0.0
